backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-appconf==1.0.5
django-cors-headers==4.3.1
django-filter==23.3
django-imagekit==5.0.0
django-nested-admin==4.0.2
django-taggit==5.0.1
djangorestframework==3.14.0
importlib-metadata==6.8.0
Markdown==3.5.1
mutagen==1.47.0
pilkit==3.0
Pillow==10.1.0
python-monkey-business==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
zipp==3.17.0
